generators/app/templates/requirements.txt
Django~=3.2.18
django-environ~=0.9.0
django-ninja~=0.20.0
django-structlog~=4.1.1
email-validator~=1.3.1
firebase-admin~=6.1.0
google-cloud-firestore~=2.9.1
google-cloud-secret-manager~=2.15.1
google-cloud-tasks~=2.12.1
gunicorn~=20.1.0
mailerlite-api-python~=0.10.0
psycopg2~=2.9.5
python-jose[cryptography]==3.3.0  # Why do we need this again?
requests~=2.28.2
